(database)/requirements.txt
# Database
supabase==2.3.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy==1.26.3

# Geospatial
geopandas==0.14.1
shapely==2.0.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0

# HTTP requests (for API extractors)
requests==2.31.0
httpx==0.26.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
